requirements.txt
anthropic==0.8.1
openai==1.6.1
langchain==0.1.0
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
packaging==23.2
semver==3.0.1
gitpython==3.1.40
pytest==7.4.3
